requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cssselect2==0.8.0
cymem==2.0.11
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.13.0
esco-skill-extractor==0.1.17
exceptiongroup==1.2.2
fastapi==0.110.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.58.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.42
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.4
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.37.2
sympy==1.13.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.29.0
waitress==3.0.2
wasabi==1.1.3
weasel==0.3.4
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
zopfli==0.2.3.post1
zstandard==0.23.0

pip~=25.1.1
setuptools~=65.5.0
langchain-google-genai~=2.1.4